requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0
python-decouple>=3.8
django-crontab>=0.7.1
Pillow>=10.0.0
reportlab>=4.0.0
weasyprint>=60.0
pytest>=7.4.0
pytest-django>=4.5.2
pytest-cov>=4.1.0
django-extensions>=3.2.3
python-magic-bin>=0.4.14; platform_system == "Windows"
python-magic>=0.4.27; platform_system != "Windows"
